downloader/requirements.txt
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
docx2python==2.8.0
et-xmlfile==1.1.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.20.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
install==1.3.5
lxml==5.2.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyDrive==1.3.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-pptx==0.6.23
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
XlsxWriter==3.2.0
yarl==1.9.4
